fix: comment함수 하나로 통일
